Controller

+++ src/main/java/com/example/simple/proj/model/User.java
@@ -1,23 +1,23 @@
 package com.example.simple.proj.model;
 
 public class User {
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private String password;
 
-    public User(int id, String name, String email, String password) {
+    public User(Long id, String name, String email, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.password = password;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
